+++ src/parser.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, jest, test } from "@jest/globals";
 import type { Parser } from "./parser";
-import { el, pure } from "./primitive";
+import { pure } from "./primitive";
 
 describe("Parser", () => {
     test("parseにArrayLike以外を入れるとエラー", () => {
@@ -19,25 +19,5 @@ describe("Parser", () => {
         const result = pure(null).flatMap(fn).parse([]);
         expect(fn).lastCalledWith(null, {});
         expect(result).toHaveProperty("value", "hoge");
-    });
-    describe("many", () => {
-        test("empty", () => {
-            expect(el(1).many().parse([])).toHaveProperty("value", []);
-        });
-        test("min", () => {
-            const parser = el(1).many({ min: 2 });
-            expect(parser.parse([1, 1])).toHaveProperty("value", [1, 1]);
-            expect(parser.parse([1, "1"])).toHaveProperty("success", false);
-            expect(parser.parse([1])).toHaveProperty("success", false);
-        });
-        test("max", () => {
-            const parser = el(1).many({ max: 2 });
-            expect(parser.parse([1, 1, 1, 1])).toHaveProperty("value", [1, 1]);
-            expect(parser.parse([1])).toHaveProperty("value", [1]);
-        });
-        test("min > max", () => {
-            const parser = el(1).many({ min: 3, max: 1 });
-            expect(parser.parse([1, 1, 1, 1])).toHaveProperty("value", [1, 1, 1]);
-        });
     });
 });