+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 ## [Unreleased]
 ### Added
 - Added `Parser.prototype.apply`
+- Added `many`
+- Added `manyAccum`
+
+### Deprecated
+- `Parser.prototype.many`
+- `Parser.prototype.manyAccum`
 
 ## [0.7.0] - 2023-05-13
 ### Added